added logic for showing checkboxes selected on refresh
