+++ src/_scss/pages/search/filters/recipient/_recipientType.scss
@@ -13,6 +13,7 @@
     font-size: rem(14);
     font-weight: 600;
     line-height: rem(21);
+    margin-left: rem(4);
   }
 
   .recipient-type-filter__item-count {
